Fix MinIOXMLParser error and refactor model definitions
